feat: add styles folder for emotion card style
