+++ YuJin/react-ts/src/app/emotions/page.tsx
@@ -1,17 +1,6 @@
 import { Link } from 'react-router-dom'
 import { EMOTION_LIST } from '../constants/emotions'
-
-type EmotionColorVariants = {
-    [key: string]: string
-}
-
-const emotionColorVariants: EmotionColorVariants = {
-    yellow: 'bg-[#FEFCE8] border border-[#fef9c2] group-hover:bg-[#fef9c2] group-hover:border-[#fff085]',
-    blue: 'bg-[#eff6ff] border border-[#dbeafe] group-hover:bg-[#dbeafe] group-hover:border-[#bedbff]',
-    green: 'bg-[#f0fdf4] border border-[#dcfce7] group-hover:bg-[#dcfce7] group-hover:border-[#b9f8cf]',
-    purple: 'bg-[#faf5ff] border border-[#f3e8ff] group-hover:bg-[#f3e8ff] group-hover:border-[#e9d4ff]',
-    red: 'bg-[#fef2f2] border border-[#ffe2e2] group-hover:bg-[#ffe2e2] group-hover:border-[#ffc9c9]',
-}
+import { emotionColorVariants } from '../../styles/emotionCard'
 export default function EmotionLinkPage() {
     return (
         <div className="flex flex-col gap-10">
